GH-1951 Extract handling to controller
